+++ js/main.js
@@ -2413,7 +2413,7 @@ function isRGBDeviceAndEnabled(device) {
     return (typeof (settings['no_rgb']) === 'undefined' ||
             (typeof (settings['no_rgb']) !== 'undefined' &&
                 parseFloat(settings['no_rgb']) === 0)) &&
-        (device['SubType'] === 'RGBW' || device['SubType'] === 'RGBWW' || device['SubType'] === 'RGB');
+        (device['SubType'] === 'RGBW' || device['SubType'] === 'RGBWW' || device['SubType'] === 'RGB' || device['SubType'] === 'RGBWWZ' );
 }
 
 function getSecurityBlock(device, idx) {