+++ js/settings.js
@@ -436,6 +436,7 @@ settingList['garbage']['garbage_company']['options']['deafvalapp'] = 'Afval App 
 settingList['garbage']['garbage_company']['options']['afvalwijzerarnhem'] = 'Afvalwijzer Arnhem (NL)';
 settingList['garbage']['garbage_company']['options']['alphenaandenrijn'] = 'Alphen aan de Rijn (NL)';
 settingList['garbage']['garbage_company']['options']['avalex'] = 'Avalex (NL)';
+settingList['garbage']['garbage_company']['options']['barafvalbeheer'] = 'Bar-afvalbeheer (Barendrecht, Rhoon)(NL)';
 settingList['garbage']['garbage_company']['options']['gemeenteberkelland'] = 'Berkelland (NL)';
 settingList['garbage']['garbage_company']['options']['best'] = 'Best (NL)';
 settingList['garbage']['garbage_company']['options']['circulusberkel'] = 'Circulus Berkel (NL)';
@@ -551,7 +552,7 @@ settingList['about']['about_text'] = {};
 settingList['about']['about_text']['title'] = 'Dashticz V' + dashticz_version + ' ' + dashticz_branch + '<br>' + newVersion;
 
 settingList['about']['about_text2'] = {};
-settingList['about']['about_text2']['title'] = 'Years after developing the old and original Dashticz, I decided to start over.<br><br>For more help visit: <a href="https://dashticz.readthedocs.io/" target="_blank">https://dashticz.readthedocs.io/</a><br>You can also check out our helpful <a href="https://www.domoticz.com/forum/viewforum.php?f=67" target="_blank">community</a> in Dashticz topic on the Domoticz forum.';
+settingList['about']['about_text2']['title'] = '<br>For more help visit: <a href="https://dashticz.readthedocs.io/" target="_blank">https://dashticz.readthedocs.io/</a><br>You can also check out our helpful <a href="https://www.domoticz.com/forum/viewforum.php?f=67" target="_blank">community</a> in Dashticz topic on the Domoticz forum.';
 
 settingList['about']['about_text4'] = {};
 settingList['about']['about_text4']['title'] = 'If you have any issues you can report them in our community thread <a href="https://www.domoticz.com/forum/viewtopic.php?f=67&t=17427" target="_blank">Bug report</a>.'