+++ docs/releasenotes/releasenotes.rst
@@ -6,6 +6,7 @@ Latest changes
 
 Fixes
   * Multiple stationclocks
+  * Background fill complete screen in case of single screen
 
 3.3.0 Beta (5-1-2020)
 ---------------------