+++ docs/releasenotes/releasenotes.rst
@@ -11,6 +11,7 @@ Fixes
   * Multiple stationclocks
   * Background fill complete screen in case of single screen
   * Add dimmer for RGBWWZ devices
+  * TwenteMilieu garbage collection
 
 3.3.0 Beta (5-1-2020)
 ---------------------